requirements.txt
flake8
ipykernel
jupyter
matplotlib
pandas
python-dotenv
scikit-learn
seaborn
